+++ integration-tests/uhppoted/tcp.py
@@ -220,6 +220,16 @@ class TestUhppoteWithTCP(unittest.TestCase):
 
         self.assertEqual(response, expected.GetCardResponse)
 
+    def test_get_card_record(self):
+        """
+        Tests the get-card-record function with defaults.
+        """
+        controller = (CONTROLLER, DEST_ADDR, "tcp")
+        card = CARD
+        record = self.u.get_card_record(controller, card)
+
+        self.assertEqual(record, expected.GetCardRecord)
+
     def test_get_card_by_index(self):
         """
         Tests the get-card-by-index function with defaults.
@@ -229,6 +239,16 @@ class TestUhppoteWithTCP(unittest.TestCase):
         response = self.u.get_card_by_index(controller, index)
 
         self.assertEqual(response, expected.GetCardByIndexResponse)
+
+    def test_get_card_record_by_index(self):
+        """
+        Tests the get-card-record-by-index function with defaults.
+        """
+        controller = (CONTROLLER, DEST_ADDR, "tcp")
+        index = CARD_INDEX
+        record = self.u.get_card_record_by_index(controller, index)
+
+        self.assertEqual(record, expected.GetCardRecordByIndex)
 
     def test_put_card(self):
         """