+++ integration-tests/uhppoted/tcp_timeout.py
@@ -208,6 +208,16 @@ class TestTCPWithTimeout(unittest.TestCase):
         self.u.get_card(controller, card)
         self.assertRaises(socket.timeout, self.u.get_card, controller, card, timeout=TIMEOUT)
 
+    def test_get_card_record(self):
+        """
+        Tests the get-card-record function with a timeout
+        """
+        controller = (CONTROLLER, DEST_ADDR, "tcp")
+        card = CARD
+
+        self.u.get_card_record(controller, card)
+        self.assertRaises(socket.timeout, self.u.get_card_record, controller, card, timeout=TIMEOUT)
+
     def test_get_card_by_index(self):
         """
         Tests the get-card-by-index function with a timeout
@@ -217,6 +227,16 @@ class TestTCPWithTimeout(unittest.TestCase):
 
         self.u.get_card_by_index(controller, index)
         self.assertRaises(socket.timeout, self.u.get_card_by_index, controller, index, timeout=TIMEOUT)
+
+    def test_get_card_record_by_index(self):
+        """
+        Tests the get-card-by-index function with a timeout
+        """
+        controller = (CONTROLLER, DEST_ADDR, "tcp")
+        index = CARD_INDEX
+
+        self.u.get_card_record_by_index(controller, index)
+        self.assertRaises(socket.timeout, self.u.get_card_record_by_index, controller, index, timeout=TIMEOUT)
 
     def test_put_card(self):
         """