+++ integration-tests/uhppoted/udp_default.py
@@ -236,6 +236,17 @@ class TestUDPWithDestAddr(unittest.TestCase):
 
         self.assertEqual(response, expected.GetCardResponse)
 
+    def test_get_card_record(self):
+        """
+        Tests the get-card-record function with defaults.
+        """
+        controller = CONTROLLER
+        card = CARD
+
+        record = self.u.get_card_record(controller, card)
+
+        self.assertEqual(record, expected.GetCardRecord)
+
     def test_get_card_by_index(self):
         """
         Tests the get-card-by-index function with defaults.
@@ -246,6 +257,17 @@ class TestUDPWithDestAddr(unittest.TestCase):
         response = self.u.get_card_by_index(controller, index)
 
         self.assertEqual(response, expected.GetCardByIndexResponse)
+
+    def test_get_card_recod_by_index(self):
+        """
+        Tests the get-card-record-by-index function with defaults.
+        """
+        controller = CONTROLLER
+        index = CARD_INDEX
+
+        record = self.u.get_card_record_by_index(controller, index)
+
+        self.assertEqual(record, expected.GetCardRecordByIndex)
 
     def test_put_card(self):
         """