+++ src/uhppoted/uhppote.py
@@ -10,6 +10,8 @@ from . import tcp
 from . import udp
 from .net import disambiguate
 
+from .structs import Card
+
 
 class Uhppote:
     """
@@ -449,6 +451,48 @@ class Uhppote:
 
         return None
 
+    def get_card_record(self, controller, card_number, timeout=2.5):
+        """
+        Retrieves the card access record for a card number from the access controller.
+            Parameters:
+               controller (uint32|tuple)  Controller serial number or tuple with (controller_id,address,protocol)
+                                          fields. The controller serial number is expected to be greater than 0.
+                                          If the controller is a tuple:
+                                          - 'controller_id' is the controller serial number
+                                          - 'address' is the optional controller IPv4 addess:port. Defaults to the
+                                             UDP broadcast address and port 60000.
+                                          - 'protocol' is an optional transport protocol ('udp' or 'tcp'). Defaults
+                                             to 'udp'.
+
+               card_number (uint32)  Access card number.
+               timeout     (float)   Optional operation timeout (in seconds). Defaults to 2.5s.
+
+            Returns:
+               Card  Card record initialised from the response from the controller.
+
+            Raises:
+               Exception  If the response from the access controller cannot be decoded.
+        """
+        (controller_id, addr, protocol) = disambiguate(controller)
+        request = encode.get_card_request(controller_id, card_number)
+
+        if reply := self._send(request, addr, timeout, protocol):
+            if response := decode.get_card_response(reply):
+                return Card(
+                    response.card_number,
+                    response.start_date,
+                    response.end_date,
+                    {
+                        1: response.door_1,
+                        2: response.door_2,
+                        3: response.door_3,
+                        4: response.door_4,
+                    },
+                    response.pin,
+                )
+
+        return None
+
     def get_card_by_index(self, controller, card_index, timeout=2.5):
         """
         Retrieves the card access record for a card record from the access controller.
@@ -477,6 +521,48 @@ class Uhppote:
 
         if reply is not None:
             return decode.get_card_by_index_response(reply)
+
+        return None
+
+    def get_card_record_by_index(self, controller, card_index, timeout=2.5):
+        """
+        Retrieves the card access record for a card record from the access controller.
+            Parameters:
+               controller (uint32|tuple)  Controller serial number or tuple with (controller_id,address,protocol)
+                                          fields. The controller serial number is expected to be greater than 0.
+                                          If the controller is a tuple:
+                                          - 'controller_id' is the controller serial number
+                                          - 'address' is the optional controller IPv4 addess:port. Defaults to the
+                                             UDP broadcast address and port 60000.
+                                          - 'protocol' is an optional transport protocol ('udp' or 'tcp'). Defaults
+                                             to 'udp'.
+
+               index       (uint32)  Controller card list record number.
+               timeout     (float)   Optional operation timeout (in seconds). Defaults to 2.5s.
+
+            Returns:
+               Card  Card record initialised from the response from the controller.
+
+            Raises:
+               Exception  If the response from the access controller cannot be decoded.
+        """
+        (controller_id, addr, protocol) = disambiguate(controller)
+        request = encode.get_card_by_index_request(controller_id, card_index)
+
+        if reply := self._send(request, addr, timeout, protocol):
+            if response := decode.get_card_by_index_response(reply):
+                return Card(
+                    response.card_number,
+                    response.start_date,
+                    response.end_date,
+                    {
+                        1: response.door_1,
+                        2: response.door_2,
+                        3: response.door_3,
+                        4: response.door_4,
+                    },
+                    response.pin,
+                )
 
         return None